requirement.txt
pandas
matplotlib
python-pptx
transformers 
requests
uvicorn
FastAPI
tensorflow
python-multipart
sentencepiece
bs4
textblob
langchain
langchain_community
langchain_text_splitters
diffusers
accelerate
scikit-learn
tf-keras
numpy
dalle-mini
pillow
